third_party/WebKit/LayoutTests/bluetooth/requestDevice/correct-uuids.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('GlucoseHeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['glucose']}],
      optionalServices: ['tx_power']}))
    .then(device => {
      assert_equals(device.uuids.length, 2);
      assert_in_array(BluetoothUUID.getService('glucose'), device.uuids);
      assert_in_array(BluetoothUUID.getService('tx_power'), device.uuids);
    });
}, 'We should only see UUID\'s that we\'ve been given permission for.')
</script>
